% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agriculture_environment}
\alias{agriculture_environment}
\title{Agriculture and environment}
\format{A data frame with 27 variables: \code{country}, \code{region},
  \code{year}, \code{}}
\usage{
agriculture_environment
}
\description{
Agriculture and environment data from gapminder.
}
\keyword{datasets}
