% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economy}
\alias{economy}
\title{Economy}
\format{An object of class \code{data.frame} with 12766 rows and 55 columns.}
\source{
https://gapminder.org/data
}
\usage{
economy
}
\description{
Economics data of African countries from Gapminder.
}
\examples{
  economy

}
\keyword{datasets}
