% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{education}
\alias{education}
\title{Education}
\format{An object of class \code{data.frame} with 2642 rows and 26 columns.}
\usage{
education
}
\description{
Education data of African countries from Gapminder
}
\keyword{datasets}
