% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health}
\alias{health}
\title{Health}
\format{An object of class \code{data.frame} with 15208 rows and 84 columns.}
\source{
https://gapminder.org/data
}
\usage{
health
}
\description{
Health data of African countries from Gapminder
}
\examples{
   health


}
\keyword{datasets}
