% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Population}
\format{An object of class \code{data.frame} with 5881 rows and 36 columns.}
\source{
https://gapminder.org/data
}
\usage{
population
}
\description{
Population data of African countries from Gapminder.
}
\examples{
   population

}
\keyword{datasets}
