% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{society}
\alias{society}
\title{Society}
\format{An object of class \code{data.frame} with 10654 rows and 45 columns.}
\source{
https://gapminder.org/data
}
\usage{
society
}
\description{
Society data of African countries from Gapminder.
}
\examples{
   society
}
\keyword{datasets}
