% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{technology}
\alias{technology}
\title{Technology}
\format{An object of class \code{data.frame} with 2397 rows and 10 columns.}
\source{
https://gapminder.org/data
}
\usage{
technology
}
\description{
Technology data of African countries from Gapminder.
}
\examples{
   technology

}
\keyword{datasets}
