% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agriculture_environment}
\alias{agriculture_environment}
\title{Agriculture and environment}
\format{An object of class \code{data.frame} with 7601 rows and 26 columns.}
\source{
https://gapminder.org/data
}
\usage{
agriculture_environment
}
\description{
Agriculture and environment data from gapminder.
}
\examples{
 agriculture_environment

}
\keyword{datasets}
