% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infrastructure}
\alias{infrastructure}
\title{Infrastructure}
\format{An object of class \code{data.frame} with 2621 rows and 9 columns.}
\usage{
infrastructure
}
\description{
Infrastructure
}
\details{
Infrastructure data of African countries from Gapminder.
}
\keyword{datasets}
