% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Population}
\format{An object of class \code{data.frame} with 8137 rows and 37 columns.}
\usage{
population
}
\description{
Population
}
\details{
Population data of African countires from Gapminder.
}
\keyword{datasets}
