% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{work}
\alias{work}
\title{Work}
\format{An object of class \code{data.frame} with 1400 rows and 38 columns.}
\usage{
work
}
\description{
Work
}
\details{
Work data of African countries from Gapminder.
}
\keyword{datasets}
